requirements.txt
urllib3==1.23
Jinja2==2.10
kubernetes==9.0.0
docker_py==1.10.6
